+++ packages/achievement/src/types/index.cairo
@@ -3,6 +3,6 @@
 #[derive(Clone, Drop, Serde, Introspect)]
 pub struct Task {
     pub id: felt252,
-    pub total: u32,
+    pub total: u128,
     pub description: ByteArray,
 }
